+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
+## [1.3.0] - 2024-09-04
+
+### Changed
+- New `NotLastByte` padding method, that is a faster version of arbitrary tail byte padding.
+- Renamed package `tests` to the more appropriate `benchmarks`.
+
 ## [1.2.0] - 2024-04-28
 
 ### Changed